Refactor Notifier class structure
